Update manage-self-service-purchases-admins.md
Below is for self-sevice sign-ups and NOT self-service purchase which is what this page is about. Edited the section to tell there is no bulk way to disable/enable.

"Use PowerShell and Microsoft Entra ID to enable or disable all self-service sign-ups
You can use PowerShell commands to change the settings that control self-service sign-ups. To turn off all self-service sign-ups, use the MSOnline PowerShell module to change the MsolCompanySettings setting for AllowAdHocSubscriptions in Microsoft Entra ID. For the steps to turn off self-service sign-ups, see Set MsolCompanySettings."