Re-remove unnecessary views from URLs
In export-topics-powershell.md, the view specified PowerShell version 7.3, but neither the reader nor the text in this article requires anything more than a version of 7 or later.

In set-up-topic-experiences.md, the view specified SharePoint PowerShell, but the other view on the target page isn't even available; trying to switch to it, returns the browser to the default view of SharePoint PowerShell.