Update loop-compliance-summary.md
updating the compliance article to lay out the information in a table format, summarize the capabilities not yet available in a separate section more succinctly (the list has shrunk!), and expand on a summary of admin management capabilities and new instructions for manual provisioning of a conditional access rule for Microsoft Entra. also, this update pushes a change to the images to include the newly released Loop components in Teams Channels, which introduces the storage of Loop components in Shared Channels. a future release of Channel meeting support will also take advantage of this new storage capability for collab notes.