Update configure-environment.md
Adjusting the phrasing to avoid confusion. Some customers thought this note means that our blob URLs do not need to be excluded from SSL inspection and they should. The point was customers should not use wildcard to exclude all blob URLs because not all blobs belong to MDE and in some cases it may be abused if wildcard exclusion is implemented.