Update m365-dr-workload-exo.md
The MS Graph PS script on lines 309-315 is not working because there is no information on how to construct the hashtable required for passing data to $LicenseDetails. I cannot find a list of attribute names or possible values, as they are not listed in https://learn.microsoft.com/en-us/powershell/module/microsoft.graph.users/update-mguserlicensedetail?view=graph-powershell-1.0.