Update microsoft-365-copilot-requirements.md
When running the new Outlook desktop app in a language that is currently not in the list of supported languages of Microsoft Copilot for Microsoft 365, the Microsoft Copilot icon will not be displayed.
Changing the Display language of Microsoft 365 to a supported language, like English, will resolve this issue. 
To reduce the likelihood of support cases, the Copilot for Microsoft 365 documentation should include this information.