Potential confusion regarding Win32 & LoB apps
Table is directly under a section named "Line-of-business apps types" in a document titled "Understand line-of-business apps for Intune". I recommend changing the wording of the section name, stripping Win32 apps from this table, or altering the wording in the Description column for Win32's row, as Win32 apps are not considered Line of Business apps. The table itself doesn't note this until the end of the description.