Update manage-creation-of-groups.md
Made some changes adding "-All" parameter to the Get-MgBetaGroup cmdlet, in order to retrieve all groups to ensure the capability to read over 100 elements in group lists. Also, I deleted the entire hash table that contains the 'EnableMSStandardBlockedWords' function since it is deprecated and retired. I only left the templateId.