Update esignature-setup.md
Added links to guest sharing both in the Prerequisites/External recipients section and the Authentication section below.   These should both align though we should discuss the fact that the content is duplicated.
I'm not sure the links were currently marked up.  The text should be 'Collaborate with guests in a site' and it should point to here https://learn.microsoft.com/en-us/microsoft-365/solutions/collaborate-in-site?view=o365-worldwide in both cases.