Update turn-bookings-on-or-off.md
"Optional: Run this command if you want to disable Bookings for all other users in your organization."  This makes the documentation sound like configuring the OWA Mailbox Policy allows for granular control of Bookings with Me, but it does not.  Corrected step 3 to clarify that this only affects the ability to create bookings calendars.