Update manage-creation-of-groups.md
Added the '-All' flag to 'Get-MgBetaGroup' command on line 140 of the file.

'Get-MgBetaGroup' only returns the first 100 results. This means the script does not work properly in environments with over 100 groups.

Adding the '-All' flag to 'Get-MgBetaGroup' allows the command to return all results, allowing the script to work properly in environments with a large volume of groups.